+++ projects/form-ui/CHANGELOG.md
@@ -1,3 +1,39 @@
+## @jaspero/fb-form-ui [1.9.1](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.0...@jaspero/fb-form-ui@1.9.1) (2021-01-07)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.13.1
+* **@jaspero/fb-tinymce:** upgraded to 1.1.8
+
+# @jaspero/fb-form-ui [1.9.0](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.8.0...@jaspero/fb-form-ui@1.9.0) (2021-01-03)
+
+
+### Features
+
+* **form-ui:** allow options to be html in select and checkbox ([3acb7c4](https://github.com/Jaspero/schema-forms/commit/3acb7c4de6dfa36e90dffbd1b6b55c795b3a101e))
+
+# @jaspero/fb-form-ui [1.8.0](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.7.3...@jaspero/fb-form-ui@1.8.0) (2020-12-28)
+
+
+### Features
+
+* **form-builder:** added not-equal option ([34a56ec](https://github.com/Jaspero/schema-forms/commit/34a56ece1d013daeb7fd437cf4b1573ac94c7a9b))
+
+## @jaspero/fb-form-ui [1.7.3](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.7.2...@jaspero/fb-form-ui@1.7.3) (2020-12-27)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.13.0
+* **@jaspero/fb-tinymce:** upgraded to 1.1.7
+
 ## @jaspero/fb-form-ui [1.7.2](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.7.1...@jaspero/fb-form-ui@1.7.2) (2020-12-23)